+++ src/components/nav/NavigationBar.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import './NavigationBar.css';
+import '../../styles/NavigationBar.css';
 
 const NavigationBar = () => {
   return (
