+++ src/pages/SettingPage.tsx
@@ -4,7 +4,7 @@ import { useAuth } from '../hooks/use-auth';
 import { removeUser } from '../store/slices/userSlice';
 import { useAppDispatch } from '../hooks/redux-hooks';
 import CategoryManager from '../components/category/CategoryManager';
-import './SettingPage.css';
+import '../styles/SettingPage.css';
 
 const SettingPage = () => {
   const dispatch = useAppDispatch();
